refactor: replace flat_zip with lszip; PackedKernel owns sole cache
Split responsibilities that were mixed in the old FlatZip class:

- lszip.h/cc: pure stateless utility — parse ZIP central directory and
  invoke a visitor callback(name, offset, size) per entry. No registry,
  no static state.

- PackedKernel::registry_ is now the sole cache, keyed
  (flatzip_path -> aks2_entry -> CachedEntry{offset, size, ptr}).
  InnerMap is fully populated from the ZIP central directory on first
  open, so inner == end is ground truth that the entry is absent.
  Fast path correctly returns nullptr for absent entries instead of
  falling through.

- CachedEntry carries offset+size (previously duplicated in
  FlatZip::registry_) alongside the PackedKernelPtr.

Adds design notes in lszip.cc on arena-based allocation strategies
considered for InnerMap population.
